+++ changelog.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [Improve ASB administrator client to return errors with specific details](https://github.com/ballerina-platform/ballerina-library/issues/8727)
 
+### Fixed
+- [Fix `updateRule` not applying filter and action changes due to incorrect field extraction from nested `SqlRule` record](https://github.com/ballerina-platform/ballerina-library/issues/8730)
+
 ## [3.9.1] - 2024-08-07
 
 #### Fixed